% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion}
\usage{
bic(Y, X, g_id, b, Delta_inv, zeta, p_thres = 0.5)
}
\arguments{
\item{Y}{response matrices; list of length \eqn{\tau} with each list component
as a \eqn{n}x\eqn{p} matrix (\eqn{p} could vary for each list
component)}

\item{X}{covariate matrix of dimension \eqn{n}x\eqn{g}}

\item{g_id}{group IDs; list of length \eqn{\tau} with each list component as
a vector of size \eqn{p} with entries \eqn{1,\ldots,m} (\eqn{p}
could vary for each list component)}

\item{b}{estimated coefficients; list of length \eqn{\tau} with each list
component as a \eqn{g}x\eqn{p} matrix (\eqn{p} could vary for each
list component)}

\item{Delta_inv}{inverse covariance matrices of the response; response
matrices; list of length \eqn{\tau} with each list component
as a \eqn{p}x\eqn{p} matrix (\eqn{p} could vary for each list
component)}

\item{zeta}{indicator variables identifying associations between response
groups and predictors; response matrices; list of length
\eqn{\tau} with each list component as a \eqn{g}x\eqn{m} matrix}

\item{p_thres}{threshold on the indicator; defaults to 0.5}
}
\value{
the computed Bayesian information criterion value
}
\description{
Computes the Bayesian information criterion from the sequential estimation
for Bayesian variable selection in multivariate multiple regression setting.
}
\examples{
n = 100
g = 10
m = 2
tau = 3
p = 5 # p_1=...=p_tau=p

Y = lapply(1:tau, function(t) matrix(rnorm(n*p), nrow = n))
X = matrix(rnorm(n*g), nrow = n)
g_id = lapply(1:tau, function(t) c(1,1,1,2,2))
b = lapply(1:tau, function(t) matrix(rnorm(g*p), nrow = g))
Delta_inv = lapply(1:tau, function(t) diag(p))
zeta = lapply(1:tau, function(t) matrix(rbinom(g*m, size=1, prob=.5), nrow=g))

bic(Y, X, g_id, b, Delta_inv, zeta)
}
