% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_gene_exp_LGG.R
\docType{data}
\name{cancer_gene_exp_LGG}
\alias{cancer_gene_exp_LGG}
\title{Gene Expression Data For LGG-Specific Cancer Driver Genes}
\format{
A matrix of dimension 63x24.
}
\usage{
cancer_gene_exp_LGG
}
\description{
Gene expression data for cancer driver genes implicated in lower
             grade gliomas (LGG). Rows correspond to the 63 LGG subjects and
             the columns correspond to the 24 LGG-specific cancer driver genes.
             Row names and column names indicate the TCGA subject ID and the
             gene name, respectively.
}
\keyword{datasets}
