% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_plot.R
\name{disc_plot}
\alias{disc_plot}
\title{Circular Disc Plots}
\usage{
disc_plot(w, im.seq, X.colnames, p_thres = 0.5, nr = 2, layer = F)
}
\arguments{
\item{w}{estimates of probability of selecting associations for the covariates
with each group of responses; list of length \eqn{\tau} with each
list component as a \eqn{g}x\eqn{m} matrix. This is usually the
output from \code{seq_est_model_selection} (accessed as
\code{seq_est_model_selection$seq_res$w}). List names should be
region IDs.}

\item{im.seq}{imaging sequence/group names; should be the same size as the number
of columns in \code{w$`Region 1`}, that is, the number of
groups in the response matrix}

\item{X.colnames}{column names of the predictor matrix (e.g. gene names)}

\item{p_thres}{threshold on probability of selection to determine inclusion
of an association; defaults to 0.5}

\item{nr}{number of rows in [ggplot2::facet_wrap]}

\item{layer}{TRUE or FALSE indicating if the region label is 'Layer' or
'Region'; Defaults to FALSE such that label is 'Region'}
}
\value{
a \code{ggplot2} object of circular disc plot
}
\description{
Circular disc plots of spherical shells indicating the selection/association
of the spherical shells with the predictors.
}
\examples{
tau = 3
m = 4
g = 20
w = lapply(1:tau, function(t) matrix(runif(m*g), nrow=g))
names(w) = paste0('Region ', 1:tau)
im.seq = paste0('IM.', 1:m)
X.colnames = LETTERS[1:g]
print(disc_plot(w, im.seq, X.colnames))
}
