% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_MBVS.R
\name{EM_MBVS}
\alias{EM_MBVS}
\title{EM Estimation For Bayesian Variable Selection}
\usage{
EM_MBVS(
  Y_t,
  X,
  g_id_t,
  b_init = NULL,
  Delta_init = NULL,
  nusq_init = NULL,
  lambda_init = NULL,
  mu = NULL,
  Lambda = NULL,
  v0 = 0.005,
  v1 = NULL,
  a1 = 4,
  a2 = 5,
  delta = NULL,
  Psi = NULL,
  maxIter = 1000,
  epsilon = 1e-05
)
}
\arguments{
\item{Y_t}{response matrix of dimension \eqn{n}x\eqn{p}}

\item{X}{covariate matrix of dimension \eqn{n}x\eqn{g}}

\item{g_id_t}{group IDs vector of size \eqn{p} with entries \eqn{1,\ldots,m}}

\item{b_init}{initial values for \eqn{\beta}; matrix of dimension
\eqn{g}x\eqn{p}. Defaults to (a) OLS estimate OR (b) OLS
estimate with g-inverse (if \eqn{X} is not full rank) OR
(c) random values sampled from normal distribution. The
default choice is made in the priority order (a),(b),(c)}

\item{Delta_init}{initial value for residual covariance \eqn{\Delta};
matrix of dimension \eqn{p}x\eqn{p}. Defaults to identity
matrix}

\item{nusq_init}{initial value for prior variance parameter of \eqn{\beta}s;
matrix of dimension \eqn{g}x\eqn{p}. Defaults to 1}

\item{lambda_init}{initial value for structure parameter; matrix of dimension
\eqn{g}x\eqn{m}. Defaults to 0}

\item{mu}{prior mean of \eqn{\lambda}; matrix of dimension \eqn{g}x\eqn{m}.
Defaults to 0}

\item{Lambda}{prior covariance of the structure parameter \eqn{\lambda};
matrix of dimension \eqn{g}x\eqn{g}. Defaults to covariance
of X}

\item{v0}{scalar factor to multiply to the variance in the prior of
\eqn{\beta}s when corresponding \eqn{\zeta = 0}. Defaults to 0.005}

\item{v1}{scalar factor to multiply to the variance in the prior of
\eqn{\beta}s when corresponding \eqn{\zeta = 1}. Defaults to
smallest power of 10 greater than the maximum magnitude of
\code{b_init}}

\item{a1}{shape parameter for Gamma prior on \eqn{nusq}. Defaults to 4}

\item{a2}{rate parameter for Gamma prior on \eqn{nusq}. Defaults to 5}

\item{delta}{degrees of freedom for the inverse-Wishart prior on
\eqn{\Delta}. Defaults to number of columns of \eqn{X}.}

\item{Psi}{prior covariance for the inverse-Wishart prior on \eqn{\Delta};
matrix of dimension \eqn{p}x\eqn{p}. Defaults to covariance
of Y_t}

\item{maxIter}{maximum number of iterations for the EM algorithm; Defaults
to 1000}

\item{epsilon}{error threshold; defaults to 1e-5}
}
\value{
\itemize{
 \item{\code{b}} {estimates of the coefficient \eqn{beta}; matrix of
                  dimension \eqn{g}x\eqn{p}}
 \item{\code{Delta_inv}} {estimate of the inverse of residual covariance;
                          matrix of dimension \eqn{p}x\eqn{p}}
 \item{\code{nusq}} {estimates of the prior variance parameter of
                     \eqn{\beta}s; matrix of dimension \eqn{g}x\eqn{p}}
 \item{\code{lambda}} {estimates of the structure parameter \eqn{\lambda};
                       matrix of dimension \eqn{g}x\eqn{m}}
 \item{\code{w}} {estimates of probability of selecting associations for
                  the covariates with each group of responses; matrix of
                  dimension \eqn{g}x\eqn{m}}
 \item{\code{iter}} {number of iterations completed by the algorithm}
 \item{\code{q_fn}} {values attained by the q-function through the
                     iterations}
 \item{\code{v1}} {scalar factor to multiply to the variance in the prior
                   of \eqn{\beta}s when corresponding \eqn{\zeta = 1}}
}
}
\description{
Expectation-Maximization based estimation for Bayesian variable selection
in multivariate multiple regression setting with grouping structure. Bayesian
variable selection can incorporate correlation structures through the priors.
}
\examples{
n = 100
g = 10
p = 5

# First three columns of Y are first group and next two columns are
# second group. Coefficients corresponding to the first three columns
# and columns 4,5 are zero and one, respectively.
g_id_t = c(1,1,1,2,2)
beta = matrix(rep(c(0,1), g*c(3,2)), nrow = g)
X = matrix(rnorm(n*g), nrow = n)
Y_t = matrix(rnorm(n*p, mean = X\%*\%beta), nrow = n)

res = EM_MBVS(Y_t, X, g_id_t)
}
