% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_prob.R
\name{inclusion_prob}
\alias{inclusion_prob}
\title{Inclusion Probabilities}
\usage{
inclusion_prob(lambda, b, g_id_t, nusq, v1, v0)
}
\arguments{
\item{lambda}{structure parameter matrix of dimension \eqn{g}x\eqn{m}}

\item{b}{coefficients matrix of dimension \eqn{g}x\eqn{p}}

\item{g_id_t}{group IDs vector of size \eqn{p} with entries \eqn{1,\ldots,m}}

\item{nusq}{prior variance parameter of \eqn{\beta}; matrix of dimension
\eqn{g}x\eqn{p}}

\item{v1}{scalar factor to multiply to the variance in the prior of
\eqn{\beta}s when corresponding \eqn{\zeta = 1}}

\item{v0}{scalar factor to multiply to the variance in the prior of
\eqn{\beta}s when corresponding \eqn{\zeta = 0}}
}
\value{
\itemize{
 \item{\code{a_prob}} {numerator of the posterior probability of the
                       indicator parameter corresponding \eqn{\zeta=1}}
 \item{\code{b_prob}} {numerator of the posterior probability of the
                       indicator parameter corresponding \eqn{\zeta=0}}
}
}
\description{
Computes the numerators of the inclusion probability, that is, the posterior
probability that the indicator parameter \eqn{\zeta} is 0 or 1.
}
\examples{
g = 10
m = 2
p = 5
v0 = 0.005
v1 = 1

# First three columns of Y are first group and next two columns are
# second group. Coefficients corresponding to the first three columns
# and columns 4,5 are zero and one, respectively.
g_id_t = c(1,1,1,2,2)
b = matrix(rep(c(0,1), g*c(3,2)), nrow = g)
nusq = matrix(rgamma(g*p, shape = 1), nrow = g)
lambda = matrix(sort(rnorm(g*m)), nrow = g)

res = inclusion_prob(lambda, b, g_id_t, nusq, v1, v0)
res$a_prob/(res$a_prob+res$b_prob)
}
