% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_on_images.R
\name{pca_on_images}
\alias{pca_on_images}
\title{Principal Component Analysis on Densities From Images}
\usage{
pca_on_images(
  image_data,
  n.regions = 3,
  den.nx = 1000,
  perc.var = 99,
  nCores = 1
)
}
\arguments{
\item{image_data}{Image data (MRI scans from various sequences) and the
segmentation mask. A list with each entry corresponding
to the imaging sequences and mask (Note: One of the
entries has to be named 'mask'). Each list entry is
again a list (of same length across all imaging sequences)
with its entries corresponding to the subjects. The
i-th list entries in each imaging sequence and the mask
correspond to subject i and are 3D-array of the same
dimensions. For example see \code{mri_data} in the package}

\item{n.regions}{number of spherical shells to divide the tumor into; defaults
to 3}

\item{den.nx}{number of points to evaluate the PDF at; defaults to 1000}

\item{perc.var}{percentage variance explained by the principal components
(PCs) to determine the number of PCs to include; defaults to
99\%}

\item{nCores}{Integer identifying the number of nodes to be forked (for
details see [parallel::makeCluster]). These nodes will be
used to run the analysis in parallel for different imaging
sequences; Defaults to 1}
}
\value{
\itemize{
\item{\code{pdfs}} {A list of size \eqn{\tau = n.regions}. Each list
                    entry is a list (size = number of imaging sequences)
                    which contains a matrix of size \eqn{n} x \eqn{den.nx}
                    of densities corresponding to an imaging sequence,
                    where \eqn{n} is the number of subjects.}
\item{\code{pc_scores}} {A list of size \eqn{\tau = n.regions}. Each list
                         entry is a matrix (size \eqn{n}x\eqn{p}) of PC
                         scores for each region. \eqn{n} is the number of
                         subjects and \eqn{p} is the number of PCs included
                         across all imaging sequences in \code{image_data}.}
\item{\code{g_id}} {A list of size \eqn{\tau = n.regions}. Each list entry
                    is a vector (size \eqn{p}) of group indicators for each
                    region, where \eqn{p} is the number of PCs included
                    across all imaging sequences in \code{image_data}.}
\item{\code{avg_pdf}} {A list of size of the number of imaging sequences in
                       \code{image_data}. Each list entry is a matrix (size
                       \eqn{n.regions} x \eqn{den.nx}) of average PDFs.}
}
}
\description{
Performs principal component analysis (PCA) on the space of probability
density functions (PDFs) generated from the tumor voxels (or other
regions of interest) from the image data. Step 1: Divide the tumor region
from each imaging sequence into \eqn{\tau} spherical shells of equal volume.
Step 2: Estimate the PDFs for each region and imaging sequence. Step 3: For
each region and each imaging sequence, compute PCA to obtain principal
component scores.
}
\examples{
n = 3
# The following will take a couple of minutes for n=3;
# Approximately 30 mins for n=63 in mri_data.

# Extracting data from first 3 subjects from MRI data
image_data = lapply(names(mri_data), function(im) mri_data[[im]][1:n])
names(image_data) = names(mri_data)

res = pca_on_images(image_data, n.regions = 3, nCores = 4)
}
