% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run data simulation}
\usage{
run_simulation(
  x,
  nc = 2000,
  ns = 3,
  nk,
  ng = nrow(x),
  de_perc,
  lfc = 2,
  c_c,
  gamma,
  beta,
  iter = 5
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{nc}{number of cells to simulate.}

\item{ns}{number of samples to simulate.}

\item{nk}{number of cell types to simulate.}

\item{ng}{number of genes to simulate.}

\item{de_perc}{percentage of cell types to be DE when initializing \deqn{X_0}.}

\item{lfc}{numeric value to use as mean logFC
(logarithm base 2) for DE genes.}

\item{c_c}{cell type relationship network matrix. 1 means connected, and 0 means not connected.}

\item{iter}{number of iterations when sampling X.}

\item{\\gamma, \\beta}{\\Phi used for sampling the state matrix X.}
}
\value{
The estimated model parameters and the DE status

a \code{\link[SingleCellExperiment]{SingleCellExperiment}}
containing multiple clusters & samples across 2 groups
as well as the following metadata: \describe{
\item{cell metadata (\code{colData(.)})}{a \code{DataFrame} containing,
containing, for each cell, it's cluster, sample, and group ID.}
\item{experiment metadata (\code{metadata(.)})}{
\describe{
\item{\code{experiment_info}}{a \code{data.frame}
summarizing the experimental design.}
\item{\code{n_cells}}{the number of cells for each sample.}
\item{\code{gene_info}}{a \code{data.frame} containing, for each gene
in each cluster, it's differential distribution \code{category},
mean \code{logFC} (NA for genes for categories "ee"),
gene used as reference (\code{sim_gene}), dispersion \code{sim_disp},
and simulation means for each group \code{sim_mean.A/B}.}
\item{\code{ref_sids/kidskids}}{the sample/cluster IDs used as reference.}
\item{\code{args}}{a list of the function call's input arguments.}}}}
}
\description{
Simulation of complex scRNA-seq data based on the sampled underlying truth X
}
