% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_est_model_selection.R
\name{seq_est_model_selection}
\alias{seq_est_model_selection}
\title{Sequential Estimation And Model Selection}
\usage{
seq_est_model_selection(
  Y,
  X,
  g_id,
  temporal = "yes",
  alpha_mix = 0.5,
  p_thres = 0.5,
  v0seq = seq(0.001, 0.01, length = 10),
  nCores = 1,
  ...
)
}
\arguments{
\item{Y}{response matrices; list of length \eqn{\tau} with each list
component as a \eqn{n}x\eqn{p} matrix (\eqn{p} could vary for
each list component)}

\item{X}{covariate matrix of dimension \eqn{n}x\eqn{g}}

\item{g_id}{group IDs; list of length \eqn{\tau} with each list component as
a vector of size \eqn{p} with entries \eqn{1,\ldots,m} (\eqn{p}
could vary for each list component)}

\item{temporal}{should the estimation borrow information sequentially ('yes'
or 'no'); Defaults to 'yes'}

\item{alpha_mix}{proportion of information to borrow; a scalar in
\eqn{[0,1]}; Defaults to 0.5}

\item{p_thres}{threshold on the indicator; defaults to 0.5}

\item{v0seq}{sequence of values to consider for v0 to perform model
estimation; Defaults to 10 equally spaced numbers between
0.001 and 0.01}

\item{nCores}{Integer identifying the number of nodes to be forked (for
details see [parallel::makeCluster]). These nodes will be
used to run estimation in parallel for different choices of
v0; Defaults to 1}

\item{...}{pass additional arguments to \code{EM_MBVS}}
}
\value{
\itemize{
 \item{\code{v0min}} {The value of v0 which has the lowest corresponding BIC}
 \item{\code{bic_res}} {A vector of values for BICs corresponding to the
                        choice of values of v0}
 \item{\code{seq_res}} {A list object for model parameter estimates with the
                        following entries:
                        \item{\code{b}} {estimates of the coefficient
                                         \eqn{beta}; list of length
                                         \eqn{\tau} with each list component
                                         as a \eqn{g}x\eqn{p} matrix
                                         (\eqn{p} could vary for each list
                                         component)}
                        \item{\code{w}} {estimates of probability of
                                         selecting associations for the
                                         covariates with each group of
                                         responses; list of length \eqn{\tau}
                                         with each list component as a
                                         \eqn{g}x\eqn{m} matrix}
                        \item{\code{lambda}} {estimates of the structure
                                              parameter \eqn{\lambda}; list
                                              of length \eqn{\tau} with each
                                              list component as a
                                              \eqn{g}x\eqn{m} matrix}
                        \item{\code{nusq}} {estimates of the prior variance
                                            parameter of \eqn{\beta}s; list
                                            of length \eqn{\tau} with each
                                            list component as a
                                            \eqn{g}x\eqn{p} matrix (\eqn{p}
                                            could vary for each list
                                            component)}
                        \item{\code{Delta_inv}} {estimate of the inverse of
                                                 residual covariance; list
                                                 of length \eqn{\tau} with
                                                 each list component as a
                                                 \eqn{p}x\eqn{p} matrix
                                                 (\eqn{p} could vary for
                                                 each list component)}
                        \item{\code{bic}} {Bayesian information criterion
                                           computed across all sequential
                                           regression model estimations; a
                                           scalar}
                        \item{\code{iter}} {number of iterations completed
                                            by the algorithm; a list with
                                            each component as an integer
                                            representing the number of EM
                                            iterations for the corresponding
                                            model in the sequence}
                        \item{\code{q_fn}} {values attained by the q-function
                                            through the iterations; a list
                                            with each component as a vector
                                            of values attained by the
                                            q-function through the iterations
                                            for the corresponding model in
                                            the sequence}
                        \item{\code{v1}} {scalar factor to multiply to the
                                          variance in the prior of
                                          \eqn{\beta}s when corresponding
                                          \eqn{\zeta = 1}}
                        }
}
}
\description{
Run sequential estimation for Bayesian variable selection in multivariate
multiple regression setting with grouping structure for a one or more
values of v0. Computes the Bayesian information criterion (BIC) from the
estimation for each value of v0. Return the parameter estimates for the
model which has the minimum BIC across different choices of v0.
}
\examples{
n = 100
g = 10
m = 2
tau = 3
p = 5 # p_1=...=p_tau=p

# First three columns of Y are first group and next two columns are
# second group. Coefficients corresponding to the first three columns
# and columns 4, 5 are zero and one, respectively. The same structure
# is maintained for t=1,...,tau
g_id = lapply(1:tau, function(t) c(1,1,1,2,2))
beta = lapply(1:tau, function(t) matrix(rep(c(0,1), g*c(3,2)), nrow = g))
X = matrix(rnorm(n*g), nrow = n)
Y = lapply(1:tau, function(t) matrix(rnorm(n*p, mean = X\%*\%beta[[t]]),
                                     nrow = n))

res = seq_est_model_selection(Y, X, g_id)
}
